pptgen/requirements.txt
python-pptx>=0.6.21
pandas>=2.0
openpyxl>=3.0
